refactor: rename outcoming_call to outbound_call and update related imports
